ML_LHP_LSTM/requirements.txt
# PyTorch and deep learning
torch>=2.0.0
torchvision>=0.15.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
SALib>=1.5.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0

# Visualization (optional, for analysis)
matplotlib>=3.7.0
seaborn>=0.12.0
